Add domain tests
whois-go doesn't allow for overriding the port, and will absolutely
call out over the wire for its data. So we have to ensure that we're
running in a privileged environment before we bind to port 43...
hence, running this test in docker.

It sucks. It's absolutely not how I want to run this stupid thing.
I might change it, depending on what CI makes available. But for now,
it at least allows for local testing.